Property_Tax_Calculation
Property Tax Calculation for Village Council is Desktop Software. This Application is use for calculating tax of properties according to their Rates and Types of building and Duration of Construction and Store the data of villagers in that village. And we have developed this project in C# .NET with Microsoft sql server.